replace git submodules with git "external"
git submodules cause developers a lot of heartache,
but are very useful when used correctly.
this combines the best features of git submodules,
with the best features of static tarball files
(and even automatically switches between the two),
 in a new feature that I'm dubbing "git external".

the main idea here is to act a bit more like svn external
(hence the name), in that this tracks a branch (or tag)
rather than a SHA1 hash.
In a nod to git submodules, however, this explicitly also remembers the
SHA1 hash of the original reference and gives a warning if they diverge.
This helps make two problem situations less likely:
  1) Pointing to a SHA1 hash that is not on an upstream branch,
     and thus later gets deleted by the git gc
  2) Forgetting to push in the correct order
     (submodule and then toplevel),
     and instantly breaking the build for everyone else

because the branch information is now stored in a file,
instead of some hidden git metadata,
this also allows building from a tarball without any extra effort.
in fact, when building from a tarball (with no .git folder),
this automatically downloads the dependency tarball as well,
so that having git installed is not necessary.

all this seamless magic is implemented in the git-external deps/Makefile rule.
see the description above that rule, and other usages, for more details.